Decouple Panel drawing and rendering.
Make it so that applications can render arbitrary content into panels by
splitting out `Pane`, a component that provides the buffer to render
into, along with input handling.

As a side effect of this refactor, Panels are now created with a
specified world-size, instead of being scaled so that the longer edge
is 1.0m.

Implements #187, #188.