Require a newer version of ndk-glue.
ndk-glue 0.6.2 moves its global state into another crate, ndk-context,
and initializes it. Other crates are starting to move to depending on
ndk-context instead, e.g. cpal depending on oboe '0.4', which is
getting resolved to a newly released 0.4.6 which depends on ndk-context.
Since we pin ndk-glue to exactly version 0.6.0, we're getting the old
ndk-glue alongside ndk-context, which remains uninitialized, leading to
explosions.

Fixes #201.